requirements.txt
Authlib[crypto]==0.3
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
